requirements.txt
#general
pip

#Gmail
google-api-python-client
google-auth
google-auth-oauthlib

#embeddings 
tiktoken
nltk
openai
chromadb
pandas
